Add all searches complete callback to itineraries search strategy api
This callback indicates there are no more searches in progress. It was necessary as client code had no way to tell when search strategy completes (update callback can be called any number of times). Resulting api is too thick, but the cost to introduce proper Rx solution is too big for this one need.